+++ stac_fastapi/types/stac_fastapi/types/rfc3339.py
@@ -6,7 +6,6 @@ from typing import Optional, Tuple, Union
 
 import iso8601
 from fastapi import HTTPException
-from pystac.utils import datetime_to_str
 
 RFC33339_PATTERN = (
     r"^(\d\d\d\d)\-(\d\d)\-(\d\d)(T|t)(\d\d):(\d\d):(\d\d)([.]\d+)?"
@@ -19,6 +18,34 @@ DateTimeType = Union[
     Tuple[datetime, None],
     Tuple[None, datetime],
 ]
+
+
+# Borrowed from pystac - https://github.com/stac-utils/pystac/blob/f5e4cf4a29b62e9ef675d4a4dac7977b09f53c8f/pystac/utils.py#L370-L394
+def datetime_to_str(dt: datetime, timespec: str = "auto") -> str:
+    """Converts a :class:`datetime.datetime` instance to an ISO8601 string in the
+    `RFC 3339, section 5.6
+    <https://datatracker.ietf.org/doc/html/rfc3339#section-5.6>`__ format required by
+    the :stac-spec:`STAC Spec <master/item-spec/common-metadata.md#date-and-time>`.
+
+    Args:
+        dt : The datetime to convert.
+        timespec: An optional argument that specifies the number of additional
+            terms of the time to include. Valid options are 'auto', 'hours',
+            'minutes', 'seconds', 'milliseconds' and 'microseconds'. The default value
+            is 'auto'.
+
+    Returns:
+        str: The ISO8601 (RFC 3339) formatted string representing the datetime.
+    """
+    if dt.tzinfo is None:
+        dt = dt.replace(tzinfo=timezone.utc)
+
+    timestamp = dt.isoformat(timespec=timespec)
+    zulu = "+00:00"
+    if timestamp.endswith(zulu):
+        timestamp = f"{timestamp[: -len(zulu)]}Z"
+
+    return timestamp
 
 
 def rfc3339_str_to_datetime(s: str) -> datetime: